+++ src/Mongo.GenericClient/Core/Repositories/IGenericRepository.cs
@@ -7,7 +7,7 @@ namespace Mongo.GenericClient.Core.Repositories
         where TEntity : IEntity
     {
         /// <summary>
-        /// The collection of <see cref="TEntity"/>.
+        /// The collection of type <see cref="TEntity"/>.
         /// </summary>
         IMongoCollection<TEntity> Collection { get; }
     }