+++ src/Mongo.GenericClient/Core/Services/IReadService.cs
@@ -12,20 +12,20 @@ namespace Mongo.GenericClient.Core.Services
         where TEntity : IEntity
     {
         /// <summary>
-        /// Get all the documents.
+        /// Get all the documents of type <see cref="TEntity"/>.
         /// </summary>
         /// <param name="filter">The filter to be applied (if not null).</param>
         /// <returns>The <see cref="IList{TEntity}"/> mapping the documents.</returns>
         IList<TEntity> GetAll(Expression<Func<TEntity, bool>>? filter = null);
 
         /// <summary>
-        /// Creates a queryable source of documents.
+        /// Creates a queryable source of documents of type <see cref="TEntity"/>.
         /// </summary>
         /// <returns>The <see cref="IMongoQueryable{TEntity}"/>.</returns>
         IMongoQueryable<TEntity> AsQueryable();
 
         /// <summary>
-        /// Get a paginated result.
+        /// Get a paginated result containing a list of type <see cref="TEntity"/>.
         /// </summary>
         /// <param name="pageNum">The page number.</param>
         /// <param name="pageSize">The page size.</param>
@@ -35,7 +35,7 @@ namespace Mongo.GenericClient.Core.Services
             int pageSize);
         
         /// <summary>
-        /// Get a paginated result.
+        /// Get a paginated result containing a list of type <see cref="TEntity"/>.
         /// </summary>
         /// <param name="filter">The filter to be applied.</param>
         /// <param name="pageNum">The page number.</param>
@@ -47,14 +47,14 @@ namespace Mongo.GenericClient.Core.Services
             int pageSize);
 
         /// <summary>
-        /// Get a document.
+        /// Get a document of type <see cref="TEntity"/>.
         /// </summary>
         /// <param name="id">The ID of the document.</param>
         /// <returns>The <see cref="TEntity"/> mapping the document.</returns>
         Task<TEntity> GetByIdAsync(ObjectId id);
 
         /// <summary>
-        /// Get a document.
+        /// Get a document of type <see cref="TEntity"/>.
         /// </summary>
         /// <param name="id">The ID of the document.</param>
         /// <returns>The <see cref="TEntity"/> mapping the document.</returns>