+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Mongo.GenericClient
 
+## Release 8.1.6
+
+- Add static class `MongoHelper`, providing access to `MongoClient` and `IMongoDatabase`
+- Minor fixes
+
 ## Release 8.1.5
 
 - Add method `DeleteAsync(string id)`
@@ -10,5 +15,5 @@
 
 ## Release 8.1.3
 
-- Add filter parameter to GetAllAsync & GetPaginated methods.
-- GetByIdAsync method: Id as ObjectId or as string.
+- Add filter parameter to `GetAllAsync` & `GetPaginated` methods
+- `GetByIdAsync` method: Id as `ObjectId` or as `string`